requirements.txt
Django==3.1.3
django-autoslug==1.9.8
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.5.0
djoser==2.1.0
drf-yasg==1.20.0
Pillow==8.0.1
pytils==0.3